Tidy v1 methods and add telemetry
