+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/ManualInstrumentation/AppSec/EventTrackingSdk/EventTrackingSdkTrackCustomEventIntegration.cs
@@ -1,8 +1,7 @@
-﻿// <copyright file="EventTrackingSdkTrackUserLoginFailureEventIntegration.cs" company="Datadog">
+﻿// <copyright file="EventTrackingSdkTrackCustomEventIntegration.cs" company="Datadog">
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
 // </copyright>
-
 #nullable enable
 
 using System.ComponentModel;
@@ -11,28 +10,28 @@ using Datadog.Trace.ClrProfiler.CallTarget;
 using Datadog.Trace.Telemetry;
 using Datadog.Trace.Telemetry.Metrics;
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.ManualInstrumentation.AppSec;
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.ManualInstrumentation.AppSec.EventTrackingSdk;
 
 /// <summary>
-/// System.Void Datadog.Trace.AppSec.EventTrackingSdk::TrackUserLoginFailureEvent(System.String,System.Boolean) calltarget instrumentation
+/// System.Void Datadog.Trace.AppSec.EventTrackingSdk::TrackCustomEvent(System.String) calltarget instrumentation
 /// </summary>
 [InstrumentMethod(
     AssemblyName = "Datadog.Trace.Manual",
     TypeName = "Datadog.Trace.AppSec.EventTrackingSdk",
-    MethodName = "TrackUserLoginFailureEvent",
+    MethodName = "TrackCustomEvent",
     ReturnTypeName = ClrNames.Void,
-    ParameterTypeNames = new[] { ClrNames.String, ClrNames.Bool },
+    ParameterTypeNames = new[] { ClrNames.String },
     MinimumVersion = ManualInstrumentationConstants.MinVersion,
     MaximumVersion = ManualInstrumentationConstants.MaxVersion,
     IntegrationName = ManualInstrumentationConstants.IntegrationName)]
 [Browsable(false)]
 [EditorBrowsable(EditorBrowsableState.Never)]
-public class EventTrackingSdkTrackUserLoginFailureEventIntegration
+public class EventTrackingSdkTrackCustomEventIntegration
 {
-    internal static CallTargetState OnMethodBegin<TTarget>(string userId, bool exists)
+    internal static CallTargetState OnMethodBegin<TTarget>(string eventName)
     {
-        TelemetryFactory.Metrics.Record(PublicApiUsage.EventTrackingSdk_TrackUserLoginFailureEvent);
-        EventTrackingSdk.TrackUserLoginFailureEvent(userId, exists, null, Datadog.Trace.Tracer.Instance);
+        TelemetryFactory.Metrics.Record(PublicApiUsage.EventTrackingSdk_TrackCustomEvent);
+        Datadog.Trace.AppSec.EventTrackingSdk.TrackCustomEvent(eventName, null, Datadog.Trace.Tracer.Instance);
         return CallTargetState.GetDefault();
     }
 }
