+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/ManualInstrumentation/AppSec/EventTrackingSdk/EventTrackingSdkTrackUserLoginFailureEventIntegration.cs
@@ -1,7 +1,8 @@
-﻿// <copyright file="EventTrackingSdkTrackCustomEventMetadataIntegration.cs" company="Datadog">
+﻿// <copyright file="EventTrackingSdkTrackUserLoginFailureEventIntegration.cs" company="Datadog">
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
 // </copyright>
+
 #nullable enable
 
 using System.ComponentModel;
@@ -10,28 +11,28 @@ using Datadog.Trace.ClrProfiler.CallTarget;
 using Datadog.Trace.Telemetry;
 using Datadog.Trace.Telemetry.Metrics;
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.ManualInstrumentation.AppSec;
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.ManualInstrumentation.AppSec.EventTrackingSdk;
 
 /// <summary>
-/// System.Void Datadog.Trace.AppSec.EventTrackingSdk::TrackCustomEvent(System.String,System.Collections.Generic.IDictionary`2[System.String,System.String]) calltarget instrumentation
+/// System.Void Datadog.Trace.AppSec.EventTrackingSdk::TrackUserLoginFailureEvent(System.String,System.Boolean) calltarget instrumentation
 /// </summary>
 [InstrumentMethod(
     AssemblyName = "Datadog.Trace.Manual",
     TypeName = "Datadog.Trace.AppSec.EventTrackingSdk",
-    MethodName = "TrackCustomEvent",
+    MethodName = "TrackUserLoginFailureEvent",
     ReturnTypeName = ClrNames.Void,
-    ParameterTypeNames = new[] { ClrNames.String, "System.Collections.Generic.IDictionary`2[System.String,System.String]" },
+    ParameterTypeNames = new[] { ClrNames.String, ClrNames.Bool },
     MinimumVersion = ManualInstrumentationConstants.MinVersion,
     MaximumVersion = ManualInstrumentationConstants.MaxVersion,
     IntegrationName = ManualInstrumentationConstants.IntegrationName)]
 [Browsable(false)]
 [EditorBrowsable(EditorBrowsableState.Never)]
-public class EventTrackingSdkTrackCustomEventMetadataIntegration
+public class EventTrackingSdkTrackUserLoginFailureEventIntegration
 {
-    internal static CallTargetState OnMethodBegin<TTarget>(string eventName, System.Collections.Generic.IDictionary<string, string> metadata)
+    internal static CallTargetState OnMethodBegin<TTarget>(string userId, bool exists)
     {
-        TelemetryFactory.Metrics.Record(PublicApiUsage.EventTrackingSdk_TrackCustomEvent_Metadata);
-        EventTrackingSdk.TrackCustomEvent(eventName, metadata, Datadog.Trace.Tracer.Instance);
+        TelemetryFactory.Metrics.Record(PublicApiUsage.EventTrackingSdk_TrackUserLoginFailureEvent);
+        Datadog.Trace.AppSec.EventTrackingSdk.TrackUserLoginFailureEvent(userId, exists, null, Datadog.Trace.Tracer.Instance);
         return CallTargetState.GetDefault();
     }
 }
