+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/ManualInstrumentation/AppSec/EventTrackingSdk/EventTrackingSdkTrackUserLoginSuccessEventIntegration.cs
@@ -8,10 +8,11 @@
 using System.ComponentModel;
 using Datadog.Trace.AppSec;
 using Datadog.Trace.ClrProfiler.CallTarget;
+using Datadog.Trace.Logging;
 using Datadog.Trace.Telemetry;
 using Datadog.Trace.Telemetry.Metrics;
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.ManualInstrumentation.AppSec;
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.ManualInstrumentation.AppSec.EventTrackingSdk;
 
 /// <summary>
 /// System.Void Datadog.Trace.AppSec.EventTrackingSdk::TrackUserLoginSuccessEvent(System.String) calltarget instrumentation
@@ -21,7 +22,7 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.ManualInstrumentation.Ap
     TypeName = "Datadog.Trace.AppSec.EventTrackingSdk",
     MethodName = "TrackUserLoginSuccessEvent",
     ReturnTypeName = ClrNames.Void,
-    ParameterTypeNames = new[] { ClrNames.String },
+    ParameterTypeNames = [ClrNames.String],
     MinimumVersion = ManualInstrumentationConstants.MinVersion,
     MaximumVersion = ManualInstrumentationConstants.MaxVersion,
     IntegrationName = ManualInstrumentationConstants.IntegrationName)]
@@ -32,7 +33,7 @@ public class EventTrackingSdkTrackUserLoginSuccessEventIntegration
     internal static CallTargetState OnMethodBegin<TTarget>(string userId)
     {
         TelemetryFactory.Metrics.Record(PublicApiUsage.EventTrackingSdk_TrackUserLoginSuccessEvent);
-        EventTrackingSdk.TrackUserLoginSuccessEvent(userId, null, Datadog.Trace.Tracer.Instance);
+        Datadog.Trace.AppSec.EventTrackingSdk.TrackUserLoginSuccessEvent(userId, null, Datadog.Trace.Tracer.Instance);
         return CallTargetState.GetDefault();
     }
 }
